Add ability to override itemStyles
Currently `itemStyles` is hardcoded and cannot be overridden.

While the options there are sensible defaults, it would be better to allow
users to override these styles when initializing the object.

Currently this can only be achieved by setting a class on the wrapper and
then overriding the CSS with an `!important` flag, and itemStyles currently
take precedent over any conflicting class styles.